acir_field
Struct GenericFieldElement

pub struct GenericFieldElement<F: PrimeField>(F);

Temporarily exported generic field to aid migration to AcirField

Tuple Fields
0: F
Implementations

impl<F: PrimeField> FieldElement<F>

pub fn from_repr(field: F) -> Self

Constructs a FieldElement from the underlying prime field representation.

This wraps an ark_ff::PrimeField element into a FieldElement.

pub fn into_repr(self) -> F

Extracts the underlying prime field representation.

This returns the wrapped ark_ff::PrimeField element.

pub fn fits_in_u128(&self) -> bool

Returns true if this field element can be represented as a u128.

A field element fits in u128 if it requires at most 128 bits to represent, i.e., if its value is in the range [0, 2^128 - 1].

pub fn fits_in_i128(&self) -> bool

Returns true if this field element can be represented as an i128.

An i128 can represent values in the range [i128::MIN, i128::MAX], which corresponds to field elements in [0, 2^127 - 1] (positive) and [p - 2^127, p - 1] (negative), where p is the field modulus. Note that 2^127 itself cannot be represented as i128 since it is not in the negative range.

pub fn try_from_str(input: &str) -> Option<FieldElement<F>>

Returns None, if the string is not a canonical representation of a field element; less than the order or if the hex string is invalid. This method can be used for both hex and decimal representations.

pub fn to_string_as_signed_integer(self, bit_size: u32) -> String

Assume this field element holds a signed integer of the given bit_size and format it as a string. The range of valid values for this field element is 0..2^bit_size with 0..2^(bit_size - 1) representing positive values and 2^(bit_size - 1)..2^bit_size representing negative values (as is commonly done for signed integers). 2^(bit_size - 1) is the lowest negative value, so for example if bit_size is 8 then 0..127 map to 0..127, 128 maps to -128, 129 maps to -127 and 255 maps to -1. If self falls outside of the valid range it’s formatted as-is.