Crate acir_field

Modules

field_element 🔒

generic_ark 🔒

Macros

assert_unique_feature

field_wrapper

Define a newtype wrapper around an AcirField by implementing all the boilerplate for forwarding the field operations.

Structs

GenericFieldElement

Temporarily exported generic field to aid migration to AcirField

Traits

AcirField

This trait is extremely unstable and WILL have breaking changes.

Functions

truncate_to

Type Aliases

FieldElement